Made some changes for testing, and got it so this can be deployed to fly.io directly.
