+++ Dockerfile
@@ -1,0 +1,6 @@
+FROM python:3.11
+WORKDIR /bot
+COPY requirements.txt /bot/
+RUN pip install -r requirements.txt
+COPY . /bot
+CMD python main.py
