--- Procfile
@@ -1,1 +1,0 @@
-worker: python bot/main.py
